Fallback for untranslated feature terms, too
This change handles the case where a language translates the `name` but
not the `terms` of a feature - the opposite of the previous commit.
Again, there would be an issue where this would previously leave `terms`
empty rather than falling back. This doesn't lead to actual crashes that
I'm aware of, but seems incorrect.