Fallback for untranslated aliases, and don't copy unnecessarily
This change handles the case where a language translates the `name` but
not the `aliases` of a feature. Note that it is not possible to translate
`aliases` without translating `name`. This means that we don't lose
information when we concatenate `name` and `aliases` in
`IDPresetsTranslationJsonParser` - the first element of `names` is always
the `name` and any subsequent elements are `aliases`.

This change also avoids unnecessary copies in
`LocalizedFeature.withFallback` when no changes would be made.

It's getting a bit unwieldy, though. There might be a bigger change to
find a simpler solution (maybe an internal data structure that
corresponds more directly (i.e. has separate `aliases`) to translations,
which gets assembled into a localized feature collection.