improve word splitter, also split in brands for search
This makes it so that the startsWith also covers the entire rest of the string, not just the single word.

Brands are not always super simple and especially brands in foreign languages might not be understood by the user as to which parts on a sign are parts of the brand or just text. People may just only try to write parts of what they can actually input on their keyboard or know how to write.

This is not perfect since a proper word split requires a lot more look-ups than just checking for characters (e.g. dictionary lookup), see https://www.unicode.org/reports/tr29/tr29-37.html

however this is much better than before and should cover more cases inside the NSI.